requirements.txt
pyserial
influxdb-client
